nearly working
